backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
openai==1.12.0
pandas==2.2.0
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24
requests==2.31.0
